Move e-mail style to a separate file
Inline CSS inside the e-mail templates is tedious – the templates are
too cluttered, the CSS needs to be synchronized between elements and
templates, adding or removing a table row means adjusting the styles
of all the rows below and there is also the issue of CSS support
in e-mail clients.

Using the emogrifier will make maintaining the templates easier.