TeamForm: do not use initialMembers when editing
When an event allows one-member teams (minMembers: 1) but the default
number of members shown by the form is higher (e.g. initialMembers: 2),
the editation would show the initial number of members instead of the
actual number of team members, when the latter was lower.

We already fixed it for form submission but editation was missed.