Fix identity handling
Nette persists identity even after logout() unless true is passed to it.

Let’s clear the identity explicitly in sign out and let’s start checking
if user is logged in before using identity just to be sure.

Also let’s log the user out if they appear to be signed in as an non-existing team.

Either of the three would prevent a bug where homepage
would not load when user whose team got deleted tried to visit it.