Switch localizations to MessageFormat
The default formatting library is too limited for our purposes.

The message catalog needs weird suffix for the values to be
considered MessageFormat so I added a custom loader.

Unfortunately, it also crashes in some weird edge cases like
https://github.com/symfony/symfony/issues/36753
so I added a test that tries to format each message.