Fix team logins not being persisted
The addition of `messages` property to `Team` introduced
a regression where the object was no longer serializable.
Let’s switch to explicitly choosing data to include
to avoid repeating this issue in the future.

Additionally, this fixes the issue where fields
like `status` would get stale in the session storage.