assets/editor: work around editor complaining about body being empty
The CodeMirror-based editor replaces textarea with a custom widget,
where user will edit the message. But for simplicity, the form still
submits the textarea content. To reduce the amount of pointless copying,
the textarea is updated when user presses a submit button.

But Nette validates the form on form submit as well. And since the Nette
code runs first, the first submit button click performed the validation
while the textarea was still empty.

Let’s define a custom pre-submit event and replace the validation function
with a one that dispatches the event before running the checks.
Then, we will be able to ensure we are validating an up-to-date form.