tests: Fix use of translation resolvers
In the last commit we explicitly enabled some resolvers in the main config,
which got merged with the tests’ config, resulting in a Session resolver
ending up being enabled in tests and tests failing due to session not existing
there. Let’s force the resolvers key in the test config to prevent that.

For simplicity, we will do that in a new config file.