Use custom Checkbox/CheckboxList/RadioList classes
Bootstrap expects that field’s error messages are a general siblings of the form entry, which has `.is-invalid` class. But that is not the case with the default Nette Forms controls.

Ideally the HTML output would be completely decoupled from the controls but until then, let’s subclass them and hack around a bit.