locale: Implement self-contained translated strings
In addition to predefined strings that are localized in translation files,
the system allows user to define new form fields in the config, where they
can be also conveniently translated.

While we could create a custom loader that would extract the translations
from the config and then use string message ids to let the `Translator`
look up translated messages in the catalogue, we would need to maintain
the loader and ensure that the message ids match and have no conflicts.

This patch chooses an easier way of using a wrapper holding all available
translations of a message, which our `Translator` will then unpack based
on the current locale.

Unfortunately, using the wrapper is not possible in some places
(e.g. for `optgroup` labels in `Nette\Forms\Controls\SelectBox`).