Support translating `optgroup` labels in `SelectBox` post-hoc
The `CategoryEntry` field is not yet attached to a form at the time it is constructed
so, unless we pass translator or locale explicitly, which is unclean and annoying,
we will need to use a value that translator can deal with so that it can
replace it at rendering time.

Unfortunately, `SelectBox` uses array keys as `optgroup` labels and PHP
does not support objects in array keys so we cannot use `NoTranslate` there.
Using a string message identifier would work but then we would have to either
declare the label translations in a translation file, or set up a custom
loader, both of which would be a pain.

Let’s instead create a custom `SelectBox` clone that supports translatable
objects as labels for `optgroup`s. We will achieve that by using `OptGroup`
objects to hold the `optgroup` information.