CustomInputModifier: delay obtaining id until form attachment
`Container::addDynamic` from `Kdyby/Replicator` would only run the closure passed to it after the form is attached.
This is not the case with `contributte/forms-multiplier`, which runs it during creation, which would result in throwing the following `InvalidStateException`:

    Component 'registry_address' is not attached to 'Nette\Forms\Form'.

Let’s wait until the form is attached before trying to obtain the input’s id.