TeamPresenter::formatData: Simplify
Use `$value` instead of `$data->$name` everywhere.

Makes the code slightly easier to read.

There is slight change in behaviour: if the data does not have the `Sportident` or `Country` field, it will be skipped, whereas, previously, it would have been listed as “to rent” or “unknown”, respectively.

But this could only occur when the configuration changed to add those fields after a team had already been registered so we do not really care.