Avoid deprecated `Identity::$id` access
`Nette\Security\SimpleIdentity` just extends the deprecated `Nette\Security\Identity` so any property access or method call on it will result in deprecation warning.

However, `User::$identity` property has `IIdentity` type hint, which is not deprecated and has `getId()` method so we can just call that and drop the cast.